requirements_test.txt
ta-cmi==3.0.0

#Test will fail witout
aiohttp_cors

pytest-homeassistant-custom-component==0.13.56

